chore: store files move to stores folder
