+++ src/components/ships/ships.tsx
@@ -6,7 +6,7 @@ import { shipsColumns } from "@/components/ships/columns.tsx";
 import { FilterComponent } from "@/components/filter/filter.tsx";
 import { units } from "@/data/units";
 import { DB_VERSION } from "@/data/version";
-import { $filterState, filterShips } from "@/filterStore";
+import { $filterState, filterShips } from "@/stores/filterStore";
 
 // This solution of wrapper component is obtained from github discussion which in turn links to discord thread
 // https://github.com/withastro/astro/issues/7709. Table from shadcn is causing issue with hydration if header
