+++ src/stores/searchStore.ts
